user-service/pyproject.toml
[tool.poetry]
name = "auth-server"
version = "0.1.0"
description = ""
authors = ["Msaleem <saleemakhtar864@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
sqlmodel = "^0.0.16"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-multipart = "^0.0.9"
types-python-jose = "^3.3.4.20240106"
types-passlib = "^1.7.7.20240327"
pytest = "^8.1.1"
pytest-asyncio = "^0.23.6"
httpx = "^0.27.0"
psycopg = {version = "^3.1.18", extras = ["binary"]}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
